parsy-backend/requirements.txt
bcrypt==3.1.7
cachetools==3.1.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
datefinder==0.7.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-SQLAlchemy==2.4.0
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
httplib2==0.13.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mysqlclient==1.4.4
oauth2client==4.1.3
oauthlib==3.1.0
pyasn1==0.4.6
pyasn1-modules==0.2.6
pycparser==2.19
python-dateutil==2.8.0
pytz==2019.2
regex==2019.8.19
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
SQLAlchemy==1.3.7
uritemplate==3.0.0
urllib3==1.25.3
Werkzeug==0.15.5
